refactor(map): split entities out and group operations under ops/
map/mod.rs had grown to 1506 lines holding the storage type, both entity
types, every operation on them, and one flat test module.

Give the entities their own modules. `MapPoint` carries real behaviour —
the median-Hamming representative descriptor pick, scale prediction, the
0.8/1.2 invariance margins — so it gets `map_point.rs`, and `Keyframe`
gets `keyframe.rs` for symmetry. The boundary is that a map point owns
what it can derive from itself; anything needing the observing keyframes
stays on `Map`, which is why `update_map_point_geometry` does not move.

Group the operations under a private `ops/` by responsibility: mutation,
correction, selection, culling, plus the existing growth and bundle
adjustment relocated unchanged. Named modules rather than a blanket
types.rs, so operation inputs and errors sit beside the code that
explains them.

Nothing about the API changes. The methods stay inherent on `Map`, the
storage fields stay private with no visibility widened, and every
`map::*` path still resolves, so callers and `use` sites are untouched.
Behaviour-preserving throughout: method bodies, validation order, epoch
increments, and locking are unchanged, and each test moved with the
operation it covers.

Verified against a pre-refactor copy: 57 method names and 29 test names
identical before and after; growth and bundle_adjustment bodies
byte-identical apart from their imports.
